sim65: add target-specific C header file and asm include file.
This PR adds a C header file (sim65.h) and an assembly include file
(sim65.inc) that define convenient access to the new sim65 peripherals
from C and assembly.

To implement these, a new export symbol "_peripherals" was added to the
linker configuration files sim6502.cfg and sim65c02.cfg, corresponding
to the base address of the peripheral memory aperture.

The C header file defines a structure rooted at _peripherals, meaning
it is possible to write things like

    peripherals.counter.latch = 0;

... and have it translated to the most efficient code possible.

The assembly include file defines a number of symbols to access registers.
In assembly, the above statement would correspond to

    lda #0
    sta peripheral_counter_latch

Apart from the addresses of the peripheral registers, both the C header
file and assembly include file also provide a bunch of constants that
are useful for readability.

Lastly, the C header file provides TRACE_ON() and TRACE_OFF() macros
that make it extremely easy to enable processor instruction tracing
from C.